Add mdb for LEO satellite. +simplify hierarchies
