--- yamcs-simulation/src/main/java/org/yamcs/simulator/PayloadInterface.java
@@ -1,10 +1,0 @@
-package org.yamcs.simulator;
-
-import java.nio.ByteBuffer;
-
-public interface PayloadInterface {
-	int getSize();
-	int getMsgId();
-	int getCRCSeed();
-	ByteBuffer makeBuffer();
-}
