--- yamcs-simulation/src/main/java/org/yamcs/simulator/UplinkInterface.java
@@ -1,9 +1,0 @@
-package org.yamcs.simulator;
-
-
-public interface UplinkInterface {
-	boolean haveAOS();
-	void sendMessageAndIgnoreLOS(PayloadInterface payload);
-	void sendMessage(PayloadInterface payload);
-	void sendMessage(PayloadInterface payload, int compid);
-}
